requirements.txt
# Flask & Extensions
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1

# AWS SDK
boto3==1.34.69
botocore==1.34.69
s3transfer==0.10.4
jmespath==1.0.1

# Environment & Utils
python-dotenv==1.1.1
Werkzeug==3.1.3
click==8.2.1
itsdangerous==2.2.0
MarkupSafe==3.0.2
greenlet==3.2.3
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
colorama==0.4.6
blinker==1.9.0
